+++ src/constants.ts
@@ -430,6 +430,26 @@ export const CONTEXT_MENU_COMMANDS = [
   COMMAND_IDS.EMOJIFY,
   COMMAND_IDS.MAKE_LONGER,
   COMMAND_IDS.MAKE_SHORTER,
+  COMMAND_IDS.TRANSLATE,
+];
+
+// Commands that process the selection
+export const PROCESS_SELECTION_COMMANDS = [
+  COMMAND_IDS.FIX_GRAMMAR,
+  COMMAND_IDS.SUMMARIZE,
+  COMMAND_IDS.GENERATE_TOC,
+  COMMAND_IDS.GENERATE_GLOSSARY,
+  COMMAND_IDS.SIMPLIFY,
+  COMMAND_IDS.EMOJIFY,
+  COMMAND_IDS.REMOVE_URLS,
+  COMMAND_IDS.REWRITE_TWEET,
+  COMMAND_IDS.REWRITE_TWEET_THREAD,
+  COMMAND_IDS.MAKE_SHORTER,
+  COMMAND_IDS.MAKE_LONGER,
+  COMMAND_IDS.ELI5,
+  COMMAND_IDS.PRESS_RELEASE,
+  COMMAND_IDS.TRANSLATE,
+  COMMAND_IDS.CHANGE_TONE,
 ];
 
 export const DEFAULT_SETTINGS: CopilotSettings = {
