+++ src/main.ts
@@ -2,22 +2,12 @@ import { BrevilabsClient } from "@/LLMProviders/brevilabsClient";
 import ChainManager from "@/LLMProviders/chainManager";
 import { CustomModel } from "@/aiParams";
 import { parseChatContent, updateChatMemory } from "@/chatUtils";
-import { addCheckCommand, addCommand, registerBuiltInCommands } from "@/commands";
+import { registerBuiltInCommands } from "@/commands";
 import CopilotView from "@/components/CopilotView";
-import { AddPromptModal } from "@/components/modals/AddPromptModal";
-import { AdhocPromptModal } from "@/components/modals/AdhocPromptModal";
-import { DebugSearchModal } from "@/components/modals/DebugSearchModal";
-import { ListPromptModal } from "@/components/modals/ListPromptModal";
 import { LoadChatHistoryModal } from "@/components/modals/LoadChatHistoryModal";
-import { OramaSearchModal } from "@/components/modals/OramaSearchModal";
-import { RemoveFromIndexModal } from "@/components/modals/RemoveFromIndexModal";
-import { SimilarNotesModal } from "@/components/modals/SimilarNotesModal";
-import { CHAT_VIEWTYPE, COMMAND_IDS, DEFAULT_OPEN_AREA, EVENT_NAMES } from "@/constants";
+import { CHAT_VIEWTYPE, DEFAULT_OPEN_AREA, EVENT_NAMES } from "@/constants";
 import { registerContextMenu } from "@/contextMenu";
-import { CustomPromptProcessor } from "@/customPromptProcessor";
 import { encryptAllKeys } from "@/encryptionService";
-import { CustomError } from "@/error";
-import { findRelevantNotes } from "@/search/findRelevantNotes";
 import { HybridRetriever } from "@/search/hybridRetriever";
 import { getAllQAMarkdownContent } from "@/search/searchUtils";
 import VectorStoreManager from "@/search/vectorStoreManager";
@@ -83,308 +73,11 @@ export default class CopilotPlugin extends Plugin {
 
     this.initActiveLeafChangeHandler();
 
-    addCommand(this, COMMAND_IDS.TOGGLE_COPILOT_CHAT_WINDOW, () => {
-      this.toggleView();
-    });
-
-    addCommand(this, COMMAND_IDS.OPEN_COPILOT_CHAT_WINDOW, () => {
-      this.activateView();
-    });
-
     this.addRibbonIcon("message-square", "Open Copilot Chat", (evt: MouseEvent) => {
       this.activateView();
     });
 
     registerBuiltInCommands(this);
-
-    const promptProcessor = CustomPromptProcessor.getInstance(this.app.vault);
-
-    addCommand(this, COMMAND_IDS.ADD_CUSTOM_PROMPT, () => {
-      new AddPromptModal(this.app, async (title: string, prompt: string) => {
-        try {
-          await promptProcessor.savePrompt(title, prompt);
-          new Notice("Custom prompt saved successfully.");
-        } catch (e) {
-          new Notice("Error saving custom prompt. Please check if the title already exists.");
-          console.error(e);
-        }
-      }).open();
-    });
-
-    addCommand(this, COMMAND_IDS.APPLY_CUSTOM_PROMPT, async () => {
-      const prompts = await promptProcessor.getAllPrompts();
-      const promptTitles = prompts.map((p) => p.title);
-      new ListPromptModal(this.app, promptTitles, async (promptTitle: string) => {
-        if (!promptTitle) {
-          new Notice("Please select a prompt title.");
-          return;
-        }
-        try {
-          const prompt = await promptProcessor.getPrompt(promptTitle);
-          if (!prompt) {
-            new Notice(`No prompt found with the title "${promptTitle}".`);
-            return;
-          }
-          this.processCustomPrompt("applyCustomPrompt", prompt.content);
-        } catch (err) {
-          console.error(err);
-          new Notice("An error occurred.");
-        }
-      }).open();
-    });
-
-    addCommand(this, COMMAND_IDS.APPLY_ADHOC_PROMPT, async () => {
-      const modal = new AdhocPromptModal(this.app, async (adhocPrompt: string) => {
-        try {
-          this.processCustomPrompt("applyAdhocPrompt", adhocPrompt);
-        } catch (err) {
-          console.error(err);
-          new Notice("An error occurred.");
-        }
-      });
-
-      modal.open();
-    });
-
-    addCheckCommand(this, COMMAND_IDS.DELETE_CUSTOM_PROMPT, (checking: boolean) => {
-      if (checking) {
-        return true;
-      }
-
-      promptProcessor.getAllPrompts().then((prompts) => {
-        const promptTitles = prompts.map((p) => p.title);
-        new ListPromptModal(this.app, promptTitles, async (promptTitle: string) => {
-          if (!promptTitle) {
-            new Notice("Please select a prompt title.");
-            return;
-          }
-
-          try {
-            await promptProcessor.deletePrompt(promptTitle);
-            new Notice(`Prompt "${promptTitle}" has been deleted.`);
-          } catch (err) {
-            console.error(err);
-            new Notice("An error occurred while deleting the prompt.");
-          }
-        }).open();
-      });
-
-      return true;
-    });
-
-    addCheckCommand(this, COMMAND_IDS.EDIT_CUSTOM_PROMPT, (checking: boolean) => {
-      if (checking) {
-        return true;
-      }
-
-      promptProcessor.getAllPrompts().then((prompts) => {
-        const promptTitles = prompts.map((p) => p.title);
-        new ListPromptModal(this.app, promptTitles, async (promptTitle: string) => {
-          if (!promptTitle) {
-            new Notice("Please select a prompt title.");
-            return;
-          }
-
-          try {
-            const prompt = await promptProcessor.getPrompt(promptTitle);
-            if (prompt) {
-              new AddPromptModal(
-                this.app,
-                async (title: string, newPrompt: string) => {
-                  try {
-                    await promptProcessor.updatePrompt(promptTitle, title, newPrompt);
-                    new Notice(`Prompt "${title}" has been updated.`);
-                  } catch (err) {
-                    console.error(err);
-                    if (err instanceof CustomError) {
-                      new Notice(err.message);
-                    } else {
-                      new Notice("An error occurred.");
-                    }
-                  }
-                },
-                prompt.title,
-                prompt.content,
-                false
-              ).open();
-            } else {
-              new Notice(`No prompt found with the title "${promptTitle}".`);
-            }
-          } catch (err) {
-            console.error(err);
-            new Notice("An error occurred.");
-          }
-        }).open();
-      });
-
-      return true;
-    });
-
-    addCommand(this, COMMAND_IDS.CLEAR_LOCAL_COPILOT_INDEX, async () => {
-      await this.vectorStoreManager.clearIndex();
-    });
-
-    addCommand(this, COMMAND_IDS.GARBAGE_COLLECT_COPILOT_INDEX, async () => {
-      try {
-        const removedDocs = await this.vectorStoreManager.garbageCollectVectorStore();
-        new Notice(`${removedDocs} documents removed from Copilot index.`);
-      } catch (err) {
-        console.error("Error garbage collecting the Copilot index:", err);
-        new Notice("An error occurred while garbage collecting the Copilot index.");
-      }
-    });
-
-    addCommand(this, COMMAND_IDS.INDEX_VAULT_TO_COPILOT_INDEX, async () => {
-      try {
-        const indexedFileCount = await this.vectorStoreManager.indexVaultToVectorStore();
-
-        new Notice(`${indexedFileCount} vault files indexed to Copilot index.`);
-        console.log(`${indexedFileCount} vault files indexed to Copilot index.`);
-      } catch (err) {
-        console.error("Error indexing vault to Copilot index:", err);
-        new Notice("An error occurred while indexing vault to Copilot index.");
-      }
-    });
-
-    addCommand(this, COMMAND_IDS.FORCE_REINDEX_VAULT_TO_COPILOT_INDEX, async () => {
-      try {
-        const indexedFileCount = await this.vectorStoreManager.indexVaultToVectorStore(true);
-
-        new Notice(`${indexedFileCount} vault files re-indexed to Copilot index.`);
-        console.log(`${indexedFileCount} vault files re-indexed to Copilot index.`);
-      } catch (err) {
-        console.error("Error re-indexing vault to Copilot index:", err);
-        new Notice("An error occurred while re-indexing vault to Copilot index.");
-      }
-    });
-
-    addCommand(this, COMMAND_IDS.LOAD_COPILOT_CHAT_CONVERSATION, () => {
-      this.loadCopilotChatHistory();
-    });
-
-    addCommand(this, COMMAND_IDS.FIND_RELEVANT_NOTES, async () => {
-      const activeFile = this.app.workspace.getActiveFile();
-      if (!activeFile) {
-        new Notice("No active file");
-        return;
-      }
-
-      const db = await this.vectorStoreManager.getDb();
-      const relevantNotes = await findRelevantNotes({
-        db,
-        filePath: activeFile.path,
-      });
-      new SimilarNotesModal(this.app, relevantNotes).open();
-    });
-
-    addCommand(this, COMMAND_IDS.INSPECT_COPILOT_INDEX_BY_NOTE_PATHS, () => {
-      new OramaSearchModal(this.app, this).open();
-    });
-
-    addCommand(this, COMMAND_IDS.SEARCH_ORAMA_DB, () => {
-      new DebugSearchModal(this.app, this).open();
-    });
-
-    addCommand(this, COMMAND_IDS.LIST_INDEXED_FILES, async () => {
-      try {
-        const indexedFiles = await this.vectorStoreManager.getIndexedFiles();
-        const indexedFilePaths = new Set(indexedFiles);
-        const allMarkdownFiles = this.app.vault.getMarkdownFiles();
-        const emptyFiles = new Set<string>();
-        const unindexedFiles = new Set<string>();
-        const filesWithoutEmbeddings = new Set<string>();
-
-        // Get dbOps for checking embeddings
-        const dbOps = await this.vectorStoreManager.getDbOps();
-
-        // Categorize files
-        for (const file of allMarkdownFiles) {
-          const content = await this.app.vault.cachedRead(file);
-          if (!content || content.trim().length === 0) {
-            emptyFiles.add(file.path);
-          } else if (!indexedFilePaths.has(file.path)) {
-            unindexedFiles.add(file.path);
-          } else {
-            // Check if file has embeddings
-            const hasEmbeddings = await dbOps.hasEmbeddings(file.path);
-            if (!hasEmbeddings) {
-              filesWithoutEmbeddings.add(file.path);
-            }
-          }
-        }
-
-        if (indexedFiles.length === 0 && emptyFiles.size === 0 && unindexedFiles.size === 0) {
-          new Notice("No files found to list.");
-          return;
-        }
-
-        // Create content for the file
-        const content = [
-          "# Copilot Files Status",
-          `- Indexed files: ${indexedFiles.length}`,
-          `	- Files missing embeddings: ${filesWithoutEmbeddings.size}`,
-          `- Unindexed files: ${unindexedFiles.size}`,
-          `- Empty files: ${emptyFiles.size}`,
-          "",
-          "## Indexed Files",
-          ...indexedFiles.map((file) => {
-            const noEmbedding = filesWithoutEmbeddings.has(file);
-            return `- [[${file}]]${noEmbedding ? " *(embedding missing)*" : ""}`;
-          }),
-          "",
-          "## Unindexed Files",
-          ...(unindexedFiles.size > 0
-            ? Array.from(unindexedFiles)
-                .sort()
-                .map((file) => `- [[${file}]]`)
-            : ["No unindexed files found."]),
-          "",
-          "## Empty Files",
-          ...(emptyFiles.size > 0
-            ? Array.from(emptyFiles)
-                .sort()
-                .map((file) => `- [[${file}]]`)
-            : ["No empty files found."]),
-        ].join("\n");
-
-        // Create or update the file in the vault
-        const fileName = `Copilot-Indexed-Files-${new Date().toLocaleDateString().replace(/\//g, "-")}.md`;
-        const filePath = `${fileName}`;
-
-        const existingFile = this.app.vault.getAbstractFileByPath(filePath);
-        if (existingFile instanceof TFile) {
-          await this.app.vault.modify(existingFile, content);
-        } else {
-          await this.app.vault.create(filePath, content);
-        }
-
-        // Open the file
-        const file = this.app.vault.getAbstractFileByPath(filePath);
-        if (file instanceof TFile) {
-          await this.app.workspace.getLeaf().openFile(file);
-          new Notice(`Listed ${indexedFiles.length} indexed files`);
-        }
-      } catch (error) {
-        console.error("Error listing indexed files:", error);
-        new Notice("Failed to list indexed files.");
-      }
-    });
-
-    addCommand(this, COMMAND_IDS.REMOVE_FILES_FROM_COPILOT_INDEX, async () => {
-      new RemoveFromIndexModal(this.app, async (filePaths: string[]) => {
-        const dbOps = await this.vectorStoreManager.getDbOps();
-        try {
-          for (const path of filePaths) {
-            await dbOps.removeDocs(path);
-          }
-          await dbOps.saveDB();
-          new Notice(`Successfully removed ${filePaths.length} files from the index.`);
-        } catch (err) {
-          console.error("Error removing files from index:", err);
-          new Notice("An error occurred while removing files from the index.");
-        }
-      }).open();
-    });
 
     this.registerEvent(
       this.app.workspace.on("editor-menu", (menu: Menu, editor: Editor) => {
